requirements.txt
discord.py>=2.3.0
python-dotenv
openai>=1.40.0
